--- a/src/makefile
+++ b/src/makefile
@@ -11,7 +11,7 @@ CFLAGS = -std=c++11 -O3 -arch=sm_35 -DDEBUG
 
 INC =
 LDFLAGS = 
-LIBS = 
+LIBS = -lcublas
 
 SOURCES = main.cu gpumd.cu read_file.cu run.cu error.cu validate.cu \
 	mic.cu atom.cu velocity.cu neighbor.cu neighbor_ON1.cu neighbor_ON2.cu \