--- a/src/modal_analysis.cuh
+++ b/src/modal_analysis.cuh
@@ -58,29 +58,18 @@ public:
     float* xdot_y;
     float* xdot_z;
 
-    float* vim_x;            // real velocity mode projection
-    float* vim_y;
-    float* vim_z;
-
-    float* vib_x;            // real velocity binned
-    float* vib_y;
-    float* vib_z;
-
     float* sqrtmass;         // precalculated mass values
     float* rsqrtmass;
 
     float* jmx;
     float* jmy;
     float* jmz;
+    float* jm;
 
     float* smx;               // stress by by square root mass
     float* smy;
     float* smz;
 
-    float* jtmp;             // placeholder for intermediate
-
-    //    float* jmn;              // per-atom modal heat current
-//    float* jm;               // total modal heat current
     float* bin_out;          // modal binning structure
     int* bin_count;          // Number of modes per bin when f_flag=1
     int* bin_sum;            // Running sum from bin_count