This work is limited to optimizing the [GKMA](#42) and [HNEMA](#43) methods. The changes are back-end and keyword usage is unchanged. I have swapped out the original custom modal analysis kernels and replaced them cuBLAS functions. The order of calculations has also been changed to reduce the number of read/write operations. Performance improvements vary depending on the size of systems, output rates, and number of bins, but initial testing has been promising. 

To give an example, an HNEMA calculation with the following conditions:
-  3680 atom graphene sheet (11040 modes)
-  output average thermal conductivity every 100 time steps 
-  52 bin output
- 10,000 step run

shows the following improvements:

|               |  Original  | Optimized | Reduction |
| --------  | --------- | ----------- | ---------- | 
| Memory | 1930 MB |   746 MB   |   ~62%    |
|    Time    |   69.4 s   |     21.6 s     |   ~69%   |  

with memory and time referring to that used during the HNEMA run in the simulation. Since the optimized version is more memory efficient, the difference between it and the original is expected to grow as the simulations scale in size.

Next, modal analysis now requires an *eigenvector.in* file as opposed to the *eigenvector.out* file used previously. There is no change to the file format, simply a change in the naming convention.

Please let me know what you think. Also, if you would like, I can try to update this code to use the GPU vectors. Otherwise, I will add those in a future pull request.

**Author(s)**
Alexander Gabourie (Stanford)