@brucefan1983 2020-05-03
This sounds very promsing. It seems now we have modified the same file during the same time and there are conflicts. Is it easy for you to resolve this?