@brucefan1983 2020-05-03
thanks. if you have any question, pleasse let me know. one change I made
yesterday is to replace LDG(a, n) to a[n]. I just know that LDG is not
needed when __restrict__ is present.

AlexG <notifications@github.com> 于 2020年5月3日周日 10:21写道：

> After merging the 'master' branch into MA_opt, my code no longer compiles.
> I will figure out the problem and fix it in the morning.
>
> —
> You are receiving this because you commented.
> Reply to this email directly, view it on GitHub
> <https://github.com/brucefan1983/GPUMD/pull/55#issuecomment-623067786>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AF546OKXRE7PXAMU6ODAJYDRPULPJANCNFSM4MX5NMFA>
> .
>