@AlexGabourie 2020-05-03
Many of the errors I'm getting are outside of my modal analysis changes. 
`error: identifier "FILE" is undefined`
is thrown for all code that uses FILE. Do you know how to fix this? Was this a problem for the current state of the master branch before I did a pull request?