@brucefan1983 2020-05-03
It might be caused by my commits after your pull? If it is not a problem
for you, you can try to create a new pull request based on the latest
master?

Bruce Fan <brucenju@gmail.com> 于 2020年5月3日周日 19:44写道：

> I think it is because I cleaned up the common.cuh file. FILE is defined in
> <stdio.h>. You can try to include this to the files that complains.
>
> Zheyong
>
>
> AlexG <notifications@github.com> 于 2020年5月3日周日 19:27写道：
>
>> Many of the errors I'm getting are outside of my modal analysis changes.
>> error: identifier "FILE" is undefined
>> is thrown for all code that uses FILE. Do you know how to fix this? Was
>> this a problem for the current state of the master branch before I did a
>> pull request?
>>
>> —
>> You are receiving this because you commented.
>> Reply to this email directly, view it on GitHub
>> <https://github.com/brucefan1983/GPUMD/pull/55#issuecomment-623136130>,
>> or unsubscribe
>> <https://github.com/notifications/unsubscribe-auth/AF546OILLIXHVXA2UPHMKPTRPWLN3ANCNFSM4MX5NMFA>
>> .
>>
>