@AlexGabourie 2020-05-03
Actually, all the commits from the master branch are in MA_opt already.  Just to double check, I downloaded a clean version of the master branch from the GPUMD GitHub page and I get the same errors. Can you get the master branch to compile?