@brucefan1983 2020-05-03
You mean you still get errors about FILE?