@brucefan1983 2020-05-03
Yes, I also got errors in Linux. 
So I will merge you pull request and fix the error.