@AlexGabourie 2020-05-03
Ok, great! Thank you! 

I think I still have some LDG usage in my code but, when I update to using GPU vectors, I will remove them.