@brucefan1983 2020-05-03
I have just added a few inclusion of <stdio.h> in a few files and it compiles in Linux now. 

You can go ahead and do the small changes you mentioned. If you have have question on the usages of GPU_Vector, just let me know.