@brucefan1983 2020-05-03
After you finished, I would like to continue to re-factor the code.